Tidy app.js, extract API function to the search component and add logo.
