+++ client/src/App.css
@@ -70,12 +70,14 @@
 }
 
 .input {
+  font-size: 1.5rem;
   width: 45vw;
   border-radius: 6px;
   border: 2px solid grey;
+  padding: 8px;
 }
 
-.submit {
+.search-button {
   border-radius: 6px;
   border: 2px solid grey;
   font-size: 1.5rem;
